Refactor "apply" code into patch package and add a test for the apply command.
