+++ cmd/registry/patch/project.go
@@ -39,7 +39,7 @@ func ExportProject(ctx context.Context, client *gapic.RegistryClient, name strin
 	}
 	err = core.ListAPIs(ctx, client, projectName.Api(""), "", func(message *rpc.Api) {
 		log.FromContext(ctx).Infof("Exporting %s", message.Name)
-		bytes, header, err := exportAPI(ctx, client, message)
+		bytes, header, err := ExportAPI(ctx, client, message)
 		if err != nil {
 			log.FromContext(ctx).WithError(err).Fatal("Failed to export artifact")
 		}
@@ -59,7 +59,7 @@ func ExportProject(ctx context.Context, client *gapic.RegistryClient, name strin
 	}
 	err = core.ListArtifacts(ctx, client, projectName.Artifact(""), "", false, func(message *rpc.Artifact) {
 		log.FromContext(ctx).Infof("Exporting %s", message.Name)
-		bytes, header, err := exportArtifact(ctx, client, message)
+		bytes, header, err := ExportArtifact(ctx, client, message)
 		if err != nil {
 			log.FromContext(ctx).WithError(err).Fatal("Failed to export artifact")
 		}
